pyproject.toml
[project]
name = "archiver"
version = "0.0.1"
authors = [
  { name = "Kaiting Chen", email = "ktchen14@gmail.com" },
]
description = "An email to JSON bridge"
readme = "README.md"
requires-python = ">= 3.11"
dependencies = [
  "Flask >= 3.0.2",
  "SQLAlchemy >= 2.0.30",
  "django-mailman3 >= 1.3.6",
  "pyjwt >= 2.8.0",
  "python-magic >= 0.4.27",

  # 3.2.0 for psycopg.Connection.notifies(timeout=...)
  "psycopg @ git+ssh://git@github.com/psycopg/psycopg.git#subdirectory=psycopg",
  "psycopg-c @ git+ssh://git@github.com/psycopg/psycopg.git#subdirectory=psycopg_c",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Flask",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Topic :: Communications :: Email",
  "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]

[project.optional-dependencies]
test = [
  "black >= 24.4.0",
  "mypy >= 1.10.0",
  "pytest >= 8.1.1",
  "ruff >= 0.4.9",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
color = true
line-length = 80

[tool.mypy]
packages = ["archiver", "test"]
strict = true
warn_no_return = false

[[tool.mypy.overrides]]
module = "test.*"
disallow_untyped_defs = false
strict_equality = false

[tool.setuptools]
packages = ["archiver"]
